requirements.txt
flask>=2.3.0
Pillow>=10.0.0
numpy>=1.24.0
scipy>=1.11.0
gunicorn>=22.0.0
